# Speed up pure Go version of Grain