The trick here is that negation turns a single bit into a mask: -0 is all zeros, while -1 is all ones.
    
The pure Go version is now faster than the assembler on amd64:
    
    name           old speed      new speed      delta
    Auth-8         88.7MB/s ± 1%  93.7MB/s ± 1%   +5.60%  (p=0.000 n=10+14)
    AuthGeneric-8  69.0MB/s ± 3%  93.5MB/s ± 1%  +35.61%  (p=0.000 n=10+13)
    Seal1K-8       44.5MB/s ± 3%  46.5MB/s ± 1%   +4.48%  (p=0.000 n=10+14)
    Open1K-8       43.8MB/s ± 1%  46.1MB/s ± 1%   +5.20%  (p=0.000 n=10+15)
    Seal8K-8       45.7MB/s ± 1%  47.5MB/s ± 1%   +3.93%  (p=0.000 n=10+15)
    Open8K-8       44.8MB/s ± 1%  47.1MB/s ± 3%   +5.18%  (p=0.000 n=10+15)
    
If the assembler version of accumulate were removed, the function could be inlined for a further speed boost.