@ericlagergren 2021-05-23
> If the assembler version of accumulate were removed, the function could be inlined for a further speed boost.

I'll look into this. I've been on the fence about the assembly implementation from the get-go because it's only marginally beneficial.